% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addflow.R
\name{addflow}
\alias{addflow}
\title{Add flow rates to a respirometry data file}
\usage{
addflow(resp, flow)
}
\arguments{
\item{resp}{Dataset from PyroScience FirestingO2 read in using read.pyro}

\item{flow}{Dataset from Respirometer's Friend read in using read.friend}
}
\value{
Returns dataset from respirometer with two columns added of flow rates
}
\description{
Will add flow rates to dataset from the PyroScience FirestingO2 using output from Respirometers Friend
}
