% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backclosed.R
\name{back.closed}
\alias{back.closed}
\title{Calculate background respiration rates from Pyro Oxygen Logger files of closed respirometery}
\usage{
back.closed(x, volume, channel = 1)
}
\arguments{
\item{x}{Dataframe from read.pyro from Pyro Oxygen Logger files}

\item{volume}{Volume in liters of the respirometer used.}

\item{channel}{The channel from which oxygen concentration data should be used.}
}
\value{
Returns background respiration rates in umolO2/min.
}
\description{
Will calculate background respiration rates for closed respirometry from the output file from the Pyro oxygen logger software read into R by the command read.pyro. This function assumes oxygen concentration was measured as umol/L.
}
