% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backpyro.R
\name{back.pyro}
\alias{back.pyro}
\title{Background respiration from Pyro Oxygen Logger Files}
\usage{
back.pyro(x, in1 = 1, in2 = 3, out1 = 2, out2 = 4, flow1, flow2)
}
\arguments{
\item{x}{Dataframe from read.pyro from Pyro Oxygen Logger files}

\item{in1}{Identifies the channel which monitors inflow of respirometer 1}

\item{in2}{Identifies the channel which monitors inflow of respirometer 2}

\item{out1}{Identifies the channel which monitors outflow of respirometer 1}

\item{out2}{Identifies the channel which monitors outflow of respirometer 2}

\item{flow1}{Flow rates through the first respirometer measured in liters per minute}

\item{flow2}{Flow rates through the second respirometer measured in liters per minute}
}
\value{
A vector of two values corresponding to the background respiration rate in first and second respirometers.
}
\description{
Will read the background respiration from a specific background run in a Pyro Oxygen Logger data file of a flow-through respirometery system. This function assumes that channels 1 and 2 are the input and output of the first respirometer, respectively, and that channels 3 & 4 are the input and output of the second respirometer, respectively.
}
