% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedresp.r
\name{closed.resp}
\alias{closed.resp}
\title{Calculate respiration rates from Pyro Oxygen Logger files of closed respirometery}
\usage{
closed.resp(x, volume, weight, channel = 1, back = 0, smooth = 0)
}
\arguments{
\item{x}{Dataframe from read.pyro from Pyro Oxygen Logger files}

\item{volume}{Volume in liters of the respirometer used.}

\item{weight}{Weight of the organism place into the respirometer measured in grams.}

\item{channel}{The channel from which oxygen concentration data should be used.}

\item{back}{The background respiration rate of the respirometer as determined from back.closed()}

\item{smooth}{The number of points over which to smooth the data.  Data points for smooth will be equall distributed before and after the point to be smoothed, and points will be weighted by a guassian distrbution with distance from point to be smoothed.}
}
\value{
Returns a vector of respiration rates in umol/g/hr.
}
\description{
Will calculate respiration rates for closed respirometry from the output file from the Pyro oxygen logger software read into R by the command read.pyro. This function assumes oxygen concentration was measured as umol/L.
}
