% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTools.R
\name{corMatrixPlot}
\alias{corMatrixPlot}
\title{Create a correlation matrix pairs plot.}
\usage{
corMatrixPlot(
  data,
  reliability,
  nullInterval = NULL,
  rscale = NULL,
  relSymbol = "r",
  draw_dist = FALSE,
  showN = FALSE,
  completedOnly = TRUE
)
}
\arguments{
\item{data}{A dataframe of data (each column is a measure, each row is a participant)}

\item{reliability}{An array of reliabilities for each measure}

\item{nullInterval}{The null interval for the correlation BF}

\item{rscale}{The scale for the correlation BF}

\item{relSymbol}{The symbol to use for reliability, can be an array of symbols for each measure}

\item{draw_dist}{A boolean indicating whether to draw the distribution on the diagonal}

\item{showN}{A boolean indicating whether to show the number of observations}

\item{completedOnly}{A boolean indicating whether to only use complete cases}
}
\value{
A ggplot object
}
\description{
Create a pairs plot for a set of measures. The diagonal typically includes
the reliabilities and a KDE of the distribution of scores across
participants. The bottom triangle plots scatter plots with a fit line. The
upper triangle indicates the correlation between the two measures with BF
and a pie chart representing the relative evidence between a positive
correlation and a point null.
}
