% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{counting}
\alias{counting}
\title{Counter function}
\usage{
counting()
}
\value{
Returns a value or vector counts of how many times the ENTER button was pressed.
}
\description{
This function will count each time the RETURN button is pressed, and start a new count each time "] followed by ENTER is pressed. Function is ended when "q" followed by ENTER is pressed. This function was programmed to work as a cell counting aid.
}
