% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTools.R
\name{disattCor}
\alias{disattCor}
\title{Calculate the disattenuated correlation}
\usage{
disattCor(rxy, rxx, ryy)
}
\arguments{
\item{rxy}{The correlation between two variables}

\item{rxx}{The reliability of the first variable}

\item{ryy}{The reliability of the second variable}
}
\value{
The disattenuated correlation
}
\description{
Calculate the disattenuated correlation
}
