% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowrate.R
\name{flowrate}
\alias{flowrate}
\title{Function to aid in the timing and calculation of gas flow rate from a volumetric pipettor flow rate apparatus}
\usage{
flowrate(increment = 1)
}
\arguments{
\item{increment}{increments (in mL) at which flow rate is to be timed}
}
\value{
Returns a dataframe containing instantaneous flow rates and cumulative flow rates
}
\description{
This function takes timed inputs (when user presses Enter) and corresponding volume increments and calculates instantaneous and cumulative flow rates based on the use of a volumetric pipette gas flow apparatus. Function is stopped when "q" is entered into the prompt.
}
