% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTools.R
\name{long2Matrix}
\alias{long2Matrix}
\title{Convert long format data to matrix (wide) format}
\usage{
long2Matrix(data, idCol, corrCol, trialNCol, summaryStats = TRUE)
}
\arguments{
\item{data}{A dataframe in the long format}

\item{idCol}{The name of the column that contains the participant id}

\item{corrCol}{The name of the column that contains the score of the trial}

\item{trialNCol}{The name of the column that contains the trial number}

\item{summaryStats}{A boolean indicating whether to add summary statistics}
}
\value{
The data in the wide format
}
\description{
This function takes a long format dataset (each row is a different
participant and a different trial) and converts it to a matrix format (each
row is a participant and each column is a trial). By default, additional
columns and rows are added with summary statistics. An additional column is
added at the end for the sum score of each participant. An additional three
rows are added for mean, variance, and item-to-rest correlation of each
trial.
}
