% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTools.R
\name{makeParcels}
\alias{makeParcels}
\title{Create parcels given a long format dataset}
\usage{
makeParcels(data, trialNCol, nParcels = 2, randomize = FALSE)
}
\arguments{
\item{data}{A dataframe in the long format}

\item{trialNCol}{The name of the column that contains the trial number}

\item{nParcels}{The number of parcels to create, default 2}

\item{randomize}{A boolean indicating whether to randomize the trials}
}
\value{
A list of tibbles with the parcels
}
\description{
This function takes a long format dataset (each row is a different
participant and a different trial) and creates parcels of the data. If the
randomize parameter is set to TRUE, the trials will be split between parcels
randomly. Otherwise, the trials will be split between parcels in order (e.g.
2 parcels will use even/odd trial numbers). The parcelsr are returned as a
list of tibbles.
}
