% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTools.R
\name{matrix2Rel}
\alias{matrix2Rel}
\title{Calculate reliability given a wide format dataset (matrix).}
\usage{
matrix2Rel(mat)
}
\arguments{
\item{mat}{A matrix of data}
}
\value{
splitHalf object
}
\description{
This function takes a a wide format dataset (matrix form) from this package,
which includes extra rows and columns for summary statistics and calculates
reliability. If the matrix does not have the extra summary statistics, you
should just use splitHalf from the psych package.
}
