% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2sat.R
\name{O2sat}
\alias{O2sat}
\title{Convert oxygen concentration data in percent atmospheric saturation into various other units}
\usage{
O2sat(value, salinity = 32, temp = 11, atmP = 1013.25)
}
\arguments{
\item{value}{Measured oxygen concentration (or vector of concentrations) of water in umolO2/L}

\item{salinity}{Salinity of water in ppt or psu. Default is 32.  Can be a vector}

\item{temp}{Temperature of water in celcius. Default is 11.  Can be a vector}

\item{atmP}{Atmospheric pressure in mbar or hPa. Default is 1013.25. Can be a vector.}
}
\value{
Returns a dataframe with one row per measured oxygen concentration value entered. Columns report salinity, temperature and atmospheric pressure used, and then retuns oxygen concentration/pressure in Saturation (%), kilopascals (kPa), hectopascals or millibar (hPa.mbar), torr, umolO2/liter, and mgO2/L.
}
\description{
This function will convert between oxygen concentration in water expressed in percent oxygen saturation to other units of concentration and pressure.
}
