% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTools-package.R
\docType{package}
\name{oTools-package}
\alias{oTools}
\alias{oTools-package}
\title{oTools: Analyze data from o tests and other similar tests from OPLab}
\description{
A set of functions useful for analyzing o tests and other similar tests created by OPLab (https://github.com/OPLabVanderbilt). Includes functions to transform data to the necessary formats and to provide a quick overview of data.
}
\author{
\strong{Maintainer}: Jason Chow \email{oTools@jasonc.how}

}
\keyword{internal}
