% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHcorr.R
\name{pHcorr}
\alias{pHcorr}
\title{Determines dye addition correction for use in spectrophotometric pH determination.}
\usage{
pHcorr(known_pH, salinity = 32, temp = 11, A434, A578, A730, B434, B578, B730)
}
\arguments{
\item{known_pH}{Known pH values of solutions. Can be a vector}

\item{salinity}{Salinity of water in ppt or psu. Default is 32.  Can be a vector}

\item{temp}{Temperature of water in celcius. Default is 11.  Can be a vector}

\item{A434}{Absorbance value of seawater + dye at 434 nm. Can be a vector.}

\item{A578}{Absorbance value of seawater + dye at 578 nm. Can be a vector.}

\item{A730}{Absorbance value of seawater + dye at 730 nm. Can be a vector.}

\item{B434}{Absorbance value of blank seawater at 434 nm. Can be a vector.}

\item{B578}{Absorbance value of blank seawater at 578 nm. Can be a vector.}

\item{B730}{Absorbance value of blank seawater at 730 nm. Can be a vector.}
}
\value{
Returns a vector of length two containing the y-intercept and coefficient of linear regression fitting the observed values to known values.
}
\description{
This function will calculate dye addition correction values to spectrophotometrically determine pH using the m-cresol purple method found in SOP 6b (Dickson, 2007).  This function take the raw absorbance values from at least two solutions of known pH (seawater pH buffer or seawater of known pCO2 and Alkalinity).
}
