% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcal.R
\name{read.cal}
\alias{read.cal}
\title{Read Calibration Data From Pyro Oxygen Logger Files}
\usage{
read.cal(file, channel)
}
\arguments{
\item{file}{File to be read}

\item{channel}{Channel for which calibration is to be read}
}
\value{
Returns an object of class "lm" based on the calibration data in file for the specified channel.
}
\description{
Will read calibration data from the output file from the Pyro oxygen logger software into R as a linear model.
}
