% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfriend.R
\name{read.friend}
\alias{read.friend}
\title{Read Respirometer's Friend Files}
\usage{
read.friend(file)
}
\arguments{
\item{file}{File to be read}
}
\value{
All data columns of the file will be read in, but header information will be discarded
}
\description{
Will read the output file from the Respirometer's Friend software into R as a dataframe.
}
