% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpresens.R
\name{read.presens}
\alias{read.presens}
\title{Read Presens Oxygen Logger Files}
\usage{
read.presens(file, convert.to.umol = T, atmP = 1013.25)
}
\arguments{
\item{file}{File to be read}

\item{convert.to.umol}{Logical. Use to disable the default behavior of converting percent oxygen saturation (what Presens records by default) to micromoles of oxygen per liter.}

\item{atmP}{Atmospheric pressure at which measurements were taken.}
}
\value{
All data columns of the file will be read in, but header information will be discarded
}
\description{
Will read the output file from the Presens oxygen logger software into R as a dataframe.
}
