% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpyro.R
\name{read.pyro}
\alias{read.pyro}
\title{Read Pyro Oxygen Logger Files}
\usage{
read.pyro(file)
}
\arguments{
\item{file}{File to be read}
}
\value{
All data columns of the file will be read in, but header information will be discarded
}
\description{
Will read the output file from the Pyro oxygen logger software into R as a dataframe.
}
