% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openresp.R
\name{resp.open}
\alias{resp.open}
\title{Calculate respiration rates from open respirometry}
\usage{
resp.open(x, inflow = 1, outflow = 2, flow.rate, weight, back = 0)
}
\arguments{
\item{x}{Dataframe from read.pyro from Pyro Oxygen Logger files}

\item{inflow}{The channel to which the inflow to the respirometer was assigned (1-4)}

\item{outflow}{The channel to which the inflow to the respirometer was assigned (1-4)}

\item{flow.rate}{Flow rates through the respirometer measured in milliliters per minute}

\item{weight}{Weight of the organism place into the respirometer measured in grams.}

\item{back}{Background respiration of the respirometer as read out by the function back.pyro}
}
\value{
Returns a data frame of four columns: times, oxy, po2 and resp. The column "times" is time in seconds. The column "oxy" is the oxygen concentration in the respirometer outflow in umolO2/L. The column "po2" is oxygen partial pressure in the respirometer outflow in kPa. The column "resp" is respiration rate in umolO2/g/hr.
}
\description{
Will calculate respiration rates for flow-through respirometry from the output file from the PyroScience FireStingO2 or Presens oxygen logger software read into R by the command read.pyro or read.presens.
}
