% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resppyro.R
\name{resp.pyro}
\alias{resp.pyro}
\title{Calculate respiration rates from Pyro Oxygen Logger files}
\usage{
resp.pyro(x, flow1, flow2, weight1, weight2, back1 = 0, back2 = 0,
  start = 0, end = max(x$times/3600))
}
\arguments{
\item{x}{Dataframe from read.pyro from Pyro Oxygen Logger files}

\item{start}{The position of the file, in hours, to begin calculating respiration.}

\item{end}{The position of the file, in hours, to end calculating respiration.}

\item{flow}{Flow rates through the first (flow1) and second (flow2) respirometers measured in liters per minute}

\item{weight}{Weight of the organism place into the first (weight1) and second (weight2) respirometers measured in grams.}

\item{back}{Background respiration of the first (back1) and second (back2) respirometers as read out by the function back.pyro}
}
\value{
Returns a data frame of four columns: oxy1 & oxy 2, the oxygen concentrations in the outflows of respirometers 1 and 2 respectively, and resp1 & resp2, respiration rates in the first and second respirometers, respectively.
}
\description{
Will calculate respiration rates for flow-through respirometry from the output file from the Pyro oxygen logger software read into R by the command read.pyro.  This function assumes that channels 1 and 2 are the input and output of the first respirometer, respectively, and that channels 3 & 4 are the input and output of the second respirometer, respectively.
}
