% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg2ppt.R
\name{sg2ppt}
\alias{sg2ppt}
\title{Convert salinity readings from a hydrometer in specific gravity to parts per thousand}
\usage{
sg2ppt(
  read,
  temp,
  cal.stand = c(25, 30, 35),
  cal.read = c(1.0185, 1.022, 1.0256),
  cal.temp = 25
)
}
\arguments{
\item{read}{Reading of the hydrometer. Can be a vector or single value.}

\item{temp}{Temperature of the water being measured in celcius. Can be a vector or single value.}

\item{cal.stand}{Vector of salinities, in ppt, of calibration solutions used.}

\item{cal.read}{Vector of hydrometer readings of calibration solutions.}

\item{cal.temp}{Vector or single value of the temperature of calibration solutions in celcius.}
}
\value{
A vector or single value of salinities in ppt calculated from the hydrometer readings in specific gravity.
}
\description{
Will convert the salinity readings of a hydrometer in specific gravity to parts per thousand. Calibration values for the hydrometer being used can be entered using the cal.stand, cal.read, & cal.temp arguments.
}
