% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specpH.R
\name{specpH}
\alias{specpH}
\title{Calculates pH of seawater from spectrophotometric data}
\usage{
specpH(salinity = 32, temp = 11, A434, A578, A730, B434, B578, B730,
  adj = c(0, 1))
}
\arguments{
\item{salinity}{Salinity of water in ppt or psu. Default is 32.  Can be a vector}

\item{temp}{Temperature of water in celcius. Default is 11.  Can be a vector}

\item{A434}{Absorbance value of seawater + dye at 434 nm. Can be a vector.}

\item{A578}{Absorbance value of seawater + dye at 578 nm. Can be a vector.}

\item{A730}{Absorbance value of seawater + dye at 730 nm. Can be a vector.}

\item{B434}{Absorbance value of blank seawater at 434 nm. Can be a vector.}

\item{B578}{Absorbance value of blank seawater at 578 nm. Can be a vector.}

\item{B730}{Absorbance value of blank seawater at 730 nm. Can be a vector.}

\item{adj}{A vector of length two containing the y-intercept and coefficient of linear regression fitting the observed values to known values.}
}
\value{
Returns a value or vector containing the calculated pH values.
}
\description{
This function will calculate the pH of seawater from the absorbance values of seawater with m-cresol purple dye. Method is based on Dickson, 2007.
}
