% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim a section out of a file using a graphic interface}
\usage{
trim(x, column)
}
\arguments{
\item{x}{Dataframe from read.pyro from Pyro Oxygen Logger files}

\item{column}{The column numebr by which you wish to select segment to be removed}
}
\value{
Returns the same object, x, with selected rows removed
}
\description{
Trim an internal segment from a an object as specified by selecting the segment in a graph.
}
