\name{OTools-package}
\alias{OTools-package}
\alias{OTools}
\docType{package}
\title{
Onthank Lab Tools
}
\description{
A collection of tools frequently used in the Onthank Cephalopod
    Ecological Physiology Lab for tasks in ecological physiology in marine
    systems.
}
\details{
\tabular{ll}{
Package: \tab OTools\cr
Type: \tab Package\cr
Version: \tab 0.21\cr
Date: \tab 2017-06-26\cr
License: \tab Creative Commons\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Kirt L Onthank

Maintainer: Kirt L Onthank <Kirt.Onthank@wallawalla.edut>

}
\references{
}
\keyword{ package }
\seealso{
}
\examples{

}
