% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resppyro.R
\name{resp.pyro}
\alias{resp.pyro}
\title{Calculate respiration rates from Pyro Oxygen Logger files}
\usage{
resp.pyro(x, flow1, flow2, weight1, weight2, back1 = 0, back2 = 0,
  start = 0, end = max(x$times/3600))
}
\arguments{
\item{x}{Dataframe from read.pyro from Pyro Oxygen Logger files}

\item{flow1}{Flow rates through the first respirometer measured in liters per minute}

\item{flow2}{Flow rates through the second respirometer measured in liters per minute}

\item{weight1}{Weight of the organism place into the first respirometer measured in grams.}

\item{weight2}{Weight of the organism place into the second respirometer measured in grams.}

\item{back1}{Background respiration of the first respirometer as read out by the function back.pyro}

\item{back2}{Background respiration of the second respirometer as read out by the function back.pyro}

\item{start}{The position of the file, in hours, to begin calculating respiration.}

\item{end}{The position of the file, in hours, to end calculating respiration.}
}
\value{
Returns a data frame of four columns: oxy1 & oxy 2, the oxygen concentrations in the outflows of respirometers 1 and 2 respectively, and resp1 & resp2, respiration rates in the first and second respirometers, respectively.
}
\description{
Will calculate respiration rates for flow-through respirometry from the output file from the Pyro oxygen logger software read into R by the command read.pyro.  This function assumes that channels 1 and 2 are the input and output of the first respirometer, respectively, and that channels 3 & 4 are the input and output of the second respirometer, respectively.
}
